handling hl update failure

+++ src/app/content/highlights/hooks/removeHighlight.ts
@@ -2,8 +2,11 @@ import { ActionHookBody } from '../../../types';
 import { actionHook } from '../../../utils';
 import { deleteHighlight } from '../actions';
 
-export const hookBody: ActionHookBody<typeof deleteHighlight> = ({highlightClient}) => async({payload}) => {
-  await highlightClient.deleteHighlight({id: payload});
-};
+export const hookBody: ActionHookBody<typeof deleteHighlight> =
+  ({highlightClient}) => async({meta, payload}) => {
+    if (meta.failedToSave) { return; }
+
+    await highlightClient.deleteHighlight({id: payload});
+  };
 
 export default actionHook(deleteHighlight, hookBody);
